requirements.txt
fastapi
uvicorn
streamlit
python-multipart
pypdf
faiss-cpu
langchain
langchain-community
langchain-google-genai
google-generativeai
python-dotenv
sentence-transformers
requests